postgres_tasks/requirements-dev.txt
amqp==5.1.1
asgiref==3.6.0
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.1.7
django-celery-beat==2.5.0
django-cors-headers==3.14.0
django-stubs==1.15.0
django-stubs-ext==0.7.0
gunicorn==20.1.0
djangorestframework==3.14.0
kombu==5.2.4
mypy==1.1.1
mypy-extensions==1.0.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.5
python-dotenv==1.0.0
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
types-pytz==2022.7.1.2
types-PyYAML==6.0.12.8
typing_extensions==4.5.0
vine==5.0.0
wcwidth==0.2.6
Pillow==9.4.0
djoser==2.1.0
django-prometheus==2.2.0
prometheus-client==0.16.0
supervisor==4.2.5